client/examples/cargo/requirements.txt
cffi==1.16.0
cryptography==42.0.8
grpcio==1.64.1
grpcio-reflection==1.64.1
grpcio-tools==1.64.1
protobuf==5.27.1
pycparser==2.22
PyJWT==2.8.0
-e git+ssh://git@github.com/hppr-dev/stoke-auth.git@d0dc5363d119736cdfed79950a086a901ee92952#egg=pystokeauth&subdirectory=client/pystokeauth
